refactor(ggw): modularize code by moving utility and LLM logic to core packages
