+++ README.md
@@ -8,3 +8,13 @@ if you have bun, give executable permission to `main.ts` and run it to run comma
 
 - [ ] setup Lib
 - [ ] Tests
+
+## RoadMap
+
+- [ ] impl `ggw`: wip
+    - [ ] tests works
+    - [ ] reproduct args: ns
+- [ ] impl `readme`: ns
+- [ ] impl `sumdiff`: ns
+- [ ] impl `which-sem`: ns
+- [ ] distribute as 1file: ns
